requirements.txt
absl-py==2.3.0
asgiref==3.8.1
astunparse==1.6.3
backports.zoneinfo==0.2.1
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.1.1
cryptography==45.0.3
cycler==0.12.1
dj-rest-auth==7.0.1
django==4.2.21
django-allauth==65.9.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
flatbuffers==25.2.10
fonttools==4.57.0
gast==0.4.0
grpcio==1.70.0
h5py==3.11.0
idna==3.10
imbalanced-learn==0.10.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
jax==0.4.13
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.1
ml-dtypes==0.2.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.4.0
packaging==25.0
pandas==2.0.3
pillow==10.4.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pymongo==4.10.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
six==1.17.0
termcolor==2.4.0
threadpoolctl==3.5.0
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
werkzeug==3.0.6
wrapt==1.14.1
zipp==3.20.2